refactor(ui)!: rename the modal-host module to modal
The runtime module behind useModal()/ModalHostProvider lives in modal.tsx
now, and the test helper moves with it (@lattice-php/ui/test/modal).

BREAKING CHANGE: deep imports move — @lattice-php/ui/modal-host is now
@lattice-php/ui/modal, and @lattice-php/ui/test/modal-host is
@lattice-php/ui/test/modal.
