+++ packages/action/resources/js/components/action.test.tsx
@@ -5,7 +5,7 @@ import { fakeNode, jsonResponse } from "@lattice-php/core/test-support";
 import { IconRendererProvider } from "@lattice-php/ui/icons";
 import type { IconRendererFunction } from "@lattice-php/ui/icons";
 import { ActionMenuProvider } from "@lattice-php/ui/action-menu-context";
-import { renderWithModalHost } from "@lattice-php/ui/test/modal-host";
+import { renderWithModalHost } from "@lattice-php/ui/test/modal";
 import ActionComponent from "./action";
 
 const apiFetch = vi.hoisted(() => vi.fn<() => Promise<Response>>());
