+++ packages/framework/resources/js/notifications/components/notifications.test.tsx
@@ -5,7 +5,7 @@ import { Renderer } from "@lattice-php/core/renderer";
 import { jsonResponse } from "@lattice-php/core/test-support";
 import type { Node } from "@lattice-php/core/types";
 import type { NotificationItem } from "@lattice-php/lattice/notifications/types";
-import { withModalHost } from "@lattice-php/ui/test/modal-host";
+import { withModalHost } from "@lattice-php/ui/test/modal";
 import { act, fireEvent, render, screen, waitFor } from "@testing-library/react";
 import type { ReactElement } from "react";
 import { describe, expect, it, vi } from "vitest";
