+++ packages/framework/resources/js/notifications/components/notifications.tsx
@@ -15,7 +15,7 @@ import {
 } from "@lattice-php/ui/primitives/popover";
 import { Icon } from "@lattice-php/ui/icons";
 import { useT } from "@lattice-php/ui/i18n";
-import { type ModalHostHandle, useModal } from "@lattice-php/ui/modal-host";
+import { type ModalHostHandle, useModal } from "@lattice-php/ui/modal";
 import type { RendererComponent } from "@lattice-php/core/types";
 import { useNotifications } from "@lattice-php/lattice/notifications/store";
 import type { NotificationItem } from "@lattice-php/lattice/notifications/types";
