+++ packages/framework/resources/js/provider-base.tsx
@@ -4,7 +4,7 @@ import type { Registry } from "@lattice-php/core/registry";
 import { RegistryProvider } from "@lattice-php/core/registry-context";
 import { useFlashEffects } from "./effects/use-flash-effects";
 import { NavigationProvider } from "@lattice-php/ui/navigation";
-import { ModalHostProvider } from "@lattice-php/ui/modal-host";
+import { ModalHostProvider } from "@lattice-php/ui/modal";
 import { EventBridge } from "./event-bridge";
 import { inertiaNavigation, useCloseModalsOnNavigate } from "./inertia-navigation";
 import type { SpriteValue } from "@lattice-php/ui/icons/sprite";
