+++ packages/media/resources/js/media-picker.test.tsx
@@ -8,7 +8,7 @@ import { useFieldScope } from "@lattice-php/form/hooks/field-scope";
 import { FormValuesProvider } from "@lattice-php/form/hooks/values";
 import { fakeNode } from "@lattice-php/core/test-support";
 import { fakeFormContext } from "@lattice-php/form/test-support";
-import { withModalHost } from "@lattice-php/ui/test/modal-host";
+import { withModalHost } from "@lattice-php/ui/test/modal";
 import { libraryRow } from "./test-support";
 import MediaPickerComponent from "./media-picker";
