+++ packages/ui/resources/js/components/modal/modal-adapter.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
 import { fakeNode } from "@lattice-php/core/test-support";
 import type { Node } from "@lattice-php/core/types";
-import { EmbeddedModalProvider } from "../../modal-host";
+import { EmbeddedModalProvider } from "../../modal";
 import ModalAdapter from "./modal-adapter";
 
 function renderModal(node: Node<"modal">, open = true, onOpenChange = vi.fn(), onExited = vi.fn()) {
