+++ packages/ui/resources/js/test/modal.tsx
@@ -1,6 +1,6 @@
 import { createElement, type ReactElement, type ReactNode } from "react";
 import { render, type RenderResult } from "@testing-library/react";
-import { ModalHostProvider } from "../modal-host";
+import { ModalHostProvider } from "../modal";
 
 /** Wraps `ui` in the host every action confirm/form overlay opens against. */
 export function withModalHost(ui: ReactNode): ReactElement {
